Change URL schema for quiz question edit/creation and add quiz deletion
