+++ backend/courses/quizzes/api/__init__.py
@@ -9,32 +9,32 @@ from courses.quizzes.api.submit_written_response_answer import (
     submit_written_response_answer,
 )
 from courses.quizzes.api.list import list_all, list_for_course
-from courses.quizzes.api.admin.api import create_quiz
-from courses.quizzes.api.admin.create_question.create_multiple_choice_question import (
+from courses.quizzes.api.admin.create_quiz import create_quiz
+from courses.quizzes.api.admin.question.create.create_multiple_choice_question import (
     create_multiple_choice_question,
 )
-from courses.quizzes.api.admin.create_question.create_checkbox_question import (
+from courses.quizzes.api.admin.question.create.create_checkbox_question import (
     create_checkbox_question,
 )
-from courses.quizzes.api.admin.create_question.create_written_response_question import (
+from courses.quizzes.api.admin.question.create.create_written_response_question import (
     create_written_response_question,
 )
-from courses.quizzes.api.admin.create_question.create_coding_question import (
+from courses.quizzes.api.admin.question.create.create_coding_question import (
     create_coding_question,
 )
-from courses.quizzes.api.admin.edit_question.edit_checkbox_question import (
+from courses.quizzes.api.admin.question.edit.edit_checkbox_question import (
     edit_checkbox_question,
 )
 
-from courses.quizzes.api.admin.edit_question.edit_coding_question import (
+from courses.quizzes.api.admin.question.edit.edit_coding_question import (
     edit_coding_question,
 )
 
-from courses.quizzes.api.admin.edit_question.edit_written_response_question import (
+from courses.quizzes.api.admin.question.edit.edit_written_response_question import (
     edit_written_response_question,
 )
 
-from courses.quizzes.api.admin.edit_question.edit_multiple_choice_question import (
+from courses.quizzes.api.admin.question.edit.edit_multiple_choice_question import (
     edit_multiple_choice_question,
 )
 
@@ -44,3 +44,5 @@ from courses.quizzes.api.admin.question.delete import (
     delete_checkbox_question,
     delete_written_response_question,
 )
+
+from courses.quizzes.api.admin.delete_quiz import delete_quiz
