+++ backend/courses/quizzes/api/admin/question/edit/edit_checkbox_question.py
