+++ backend/tests/courses/quizzes/admin/question/create/test_create_coding_question.py
@@ -5,7 +5,7 @@ from rest_framework import status
 
 class CreateCodingQuestionTests(TestCasesWithUserAuth):
     def get_api_endpoint(self, course_slug: str, quiz_slug: str) -> str:
-        return f'/api/v0/quizzes/admin/{course_slug}/{quiz_slug}/create/coding/'
+        return f'/api/v0/quizzes/admin/{course_slug}/{quiz_slug}/coding/create/'
 
     def test_happy_path(self):
         # Create a quiz and enroll the user as an instructor
